backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.2.0
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
